+++ core-parent/admin-core-dto/src/main/java/com/kalix/admin/core/dto/model/FunctionDTO.java
@@ -17,9 +17,9 @@ public class FunctionDTO extends BaseDTO {
     private String code; //功能代码
     private String text; //名称
     private boolean leaf; //是否是叶子节点
-    private long parentId; //父功能
+    private Long parentId; //父功能
     private String remark; //备注
-    private long applicationId;  // 归属应用
+    private Long applicationId;  // 归属应用
     private String parentName; //父节点名称
     private String permission; //权限路径
     private List<FunctionDTO> children=new ArrayList<FunctionDTO>();
@@ -81,19 +81,19 @@ public class FunctionDTO extends BaseDTO {
         this.leaf = leaf;
     }
 
-    public long getParentId() {
+    public Long getParentId() {
         return parentId;
     }
 
-    public void setParentId(long parentId) {
+    public void setParentId(Long parentId) {
         this.parentId = parentId;
     }
 
-    public long getApplicationId() {
+    public Long getApplicationId() {
         return applicationId;
     }
 
-    public void setApplicationId(int applicationId) {
+    public void setApplicationId(Long applicationId) {
         this.applicationId = applicationId;
     }
