+++ core-parent/admin-core-entities/src/main/java/com/kalix/admin/core/entities/MessageBean.java
@@ -21,8 +21,8 @@ public class MessageBean extends PersistentEntity {
     private String sender = "系统管理员";//发送人
     private String title;//消息名称
     private String message;//消息内容
-    private long userId;//接收用户名称
-    private long status = 0;//消息状态，0为未读，1为已读
+    private Long userId;//接收用户名称
+    private Long status = 0L;//消息状态，0为未读，1为已读
 
     public String getSender() {
         return sender;
@@ -40,19 +40,19 @@ public class MessageBean extends PersistentEntity {
         this.message = message;
     }
 
-    public long getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(long userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
-    public long getStatus() {
+    public Long getStatus() {
         return status;
     }
 
-    public void setStatus(int status) {
+    public void setStatus(Long status) {
         this.status = status;
     }
